Module wasmtime_fuzzing::generators

Test case generators.

Test case generators take raw, unstructured input from a fuzzer (e.g. libFuzzer) and translate that into a structured test case (e.g. a valid Wasm binary).

These are generally implementations of the Arbitrary trait, or some wrapper over an external tool, such that the wrapper implements the Arbitrary trait for the wrapped external tool.

Modules

api

Generating sequences of Wasmtime API calls.

table_ops

Generating series of table.get and table.set operations.

Structs

Config

Implementation of generating a wasmtime::Config arbitrarily

DifferentialConfig

A description of configuration options that we should do differential testing between.

SpecTest

A spec test from the upstream wast testsuite, arbitrarily chosen from the list of known spec tests.

WasmtimeDefaultConfig

Wasmtime-specific default configuration for wasm-smith-generated modules.

Type Definitions

GeneratedModule

Type alias for wasm-smith generated modules using wasmtime’s default configuration.