Module wasmtime_fuzzing::generators::api

Generating sequences of Wasmtime API calls.

We only generate valid sequences of API calls. To do this, we keep track of what objects we’ve already created in earlier API calls via the Scope struct.

To generate even-more-pathological sequences of API calls, we use swarm testing:

In swarm testing, the usual practice of potentially including all features in every test case is abandoned. Rather, a large “swarm” of randomly generated configurations, each of which omits some features, is used, with configurations receiving equal resources.

Structs

ApiCalls

A sequence of API calls.

Enums

ApiCall

A call to one of Wasmtime’s public APIs.