Module wasmtime_fuzzing::oracles

Oracles.

Oracles take a test case and determine whether we have a bug. For example, one of the simplest oracles is to take a Wasm binary as our input test case, validate and instantiate it, and (implicitly) check that no assertions failed or segfaults happened. A more complicated oracle might compare the result of executing a Wasm file with and without optimizations enabled, and make sure that the two executions are observably identical.

When an oracle finds a bug, it should report it to the fuzzing engine by panicking.

Modules

dummy

Dummy implementations of things that a Wasm module can import.

Structs

SingleFunctionModuleConfig

Configuration options for wasm-smith such that generated modules always conform to certain specifications: one exported function, one exported memory.

Enums

Timeout

Methods of timing out execution of a WebAssembly module

Functions

compile

Compile the Wasm buffer, and implicitly fail if we have an unexpected panic or segfault or anything else that can be detected “passively”.

differential_execution

Instantiate the given Wasm module with each Config and call all of its exports. Modulo OOM, non-canonical NaNs, and usage of Wasm features that are or aren’t enabled for different configs, we should get the same results when we call the exported functions for all of our different configs.

differential_spec_execution

Perform differential execution between Wasmtime and the official WebAssembly specification interpreter.

differential_v8_execution

Performs differential execution between Wasmtime and V8.

differential_wasmi_execution

Perform differential execution between Cranelift and wasmi, diffing the resulting memory image when execution terminates. This relies on the module-under-test to be instrumented to bound the execution time. Invoke with a module generated by wasm-smith using the SingleFunctionModuleConfig configuration type for best results.

instantiate

Instantiate the Wasm buffer, and implicitly fail if we have an unexpected panic or segfault or anything else that can be detected “passively”.

instantiate_with_config

Instantiate the Wasm buffer, and implicitly fail if we have an unexpected panic or segfault or anything else that can be detected “passively”.

make_api_calls

Invoke the given API calls.

spectest

Executes the wast test spectest with the config specified.

table_ops

Execute a series of table.get and table.set operations.