Module wasmtime_fuzzing::oracles::dummy

Dummy implementations of things that a Wasm module can import.

Functions

dummy_extern
Construct a dummy Extern from its type signature

dummy_func
Construct a dummy function for the given function type

dummy_global
Construct a dummy global for the given global type.

dummy_instance
Construct a dummy instance for the given instance type.

dummy_linker
Create a set of dummy functions/globals/etc for the given imports.

dummy_memory
Construct a dummy memory for the given memory type.

dummy_module
Construct a dummy module for the given module type.

dummy_table
Construct a dummy table for the given table type.

dummy_value
Construct a dummy value for the given value type.

dummy_values
Construct a sequence of dummy values for the given types.