Module wasmtime_fuzzing::oracles::engine

Define the interface for differential evaluation of Wasm functions.

Traits

DiffEngine

Provide a way to instantiate Wasm modules.

DiffInstance

Provide a way to evaluate Wasm functions–a Wasm instance implemented by a specific engine (i.e., compiler or interpreter).

Functions

build

Returns a function which can be used to build the engine name specified.

build_allowed_env_list

Build a list of allowed values from the given defaults using the env_list.

parse_env_list

Retrieve a comma-delimited list of values from an environment variable.

setup_engine_runtimes

Initialize any global state associated with runtimes that may be differentially executed against.