Enum wasmtime_fuzzing::oracles::DiffEqResult

pub enum DiffEqResult<T, U> {
    Success(T, U),
    Poisoned,
    Failed,
}

Result of comparing the result of two operations during differential execution.

Variants

Success(T, U)

Both engines succeeded.

Poisoned

The result has reached the state where engines may have diverged and results can no longer be compared.

Failed

Both engines failed with the same error message, and internal state should still match between the two engines.

Implementations

impl<T, U> DiffEqResult<T, U>

pub fn new( lhs_engine: &dyn DiffEngine, lhs_result: Result<T>, rhs_result: Result<U> ) -> DiffEqResult<T, U>

Computes the differential result from executing in two different engines.