Function wasmtime_fuzzing::oracles::differential

pub fn differential(
    lhs: &mut dyn DiffInstance, 
    lhs_engine: &dyn DiffEngine, 
    rhs: &mut WasmtimeInstance, 
    name: &str, 
    args: &[DiffValue], 
    result_tys: &[DiffValueType]
) -> Result<bool>

Evaluate the function identified by name in two different engine instances–lhs and rhs.

Returns Ok(true) if more evaluations can happen or Ok(false) if the instances may have drifted apart and no more evaluations can happen.

Panics

This will panic if the evaluation is different between engines (e.g., results are different, hashed instance is different, one side traps, etc.).