Function wasmtime_fuzzing::oracles::spectest

pub fn spectest(config: Config, test: SpecTest)

Executes the wast test spectest with the config specified.

Ensures that spec tests pass regardless of the Config.