Function wasmtime_fuzzing::single_module_fuzzer::mutate

pub fn mutate<T>(
    data: &mut [u8],
    size: usize,
    max_size: usize,
    gen_module: fn(_: &mut T, _: &mut Unstructured<'_>) -> Result<(Vec<u8>, KnownValid)>,
    mutate: fn(_: &mut [u8], _: usize, _: usize) -> usize
) -> usize
where
    T: for<'a> Arbitrary<'a>,

Implementation of a libfuzzer custom mutator for a single-module-fuzzer.

This mutator will take the seed specified in data and attempt to mutate it with the provided mutate function. The mutate function may not receive the data as-specified, but instead may receive only the seed that was used to generate data.