Module wasmtime_fuzzing::oracles

Oracles.

Oracles take a test case and determine whether we have a bug. For example, one of the simplest oracles is to take a Wasm binary as our input test case, validate and instantiate it, and (implicitly) check that no assertions failed or segfaults happened. A more complicated oracle might compare the result of executing a Wasm file with and without optimizations enabled, and make sure that the two executions are observably identical.

When an oracle finds a bug, it should report it to the fuzzing engine by panicking.

Modules

dummy
Dummy implementations of things that a Wasm module can import.

Functions

compile
Compile the Wasm buffer, and implicitly fail if we have an unexpected panic or segfault or anything else that can be detected "passively".

instantiate
Instantiate the Wasm buffer, and implicitly fail if we have an unexpected panic or segfault or anything else that can be detected "passively".

instantiate_with_config
Instantiate the Wasm buffer, and implicitly fail if we have an unexpected panic or segfault or anything else that can be detected "passively".

spectest
Executes the wast test spectest with the config specified.